# openapi.yaml
openapi: 3.1.0
info:
  title: attn-viewer API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /api/blogs:
    get:
      summary: List registered Naver blogs and collection status
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Blog registry with corpus statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Registry read failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Register or reactivate a Naver blog and start its initial crawl
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogCreateRequest"
      responses:
        "202":
          description: Blog accepted and crawl job requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogCreateResponse"
        "400":
          description: Invalid blog id or crawl start time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: The blog could not be verified or has no public posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/blogs/posts:
    get:
      summary: List posts from all active registered blogs
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/BlogPostOffset"
        - $ref: "#/components/parameters/BlogPostLimit"
      responses:
        "200":
          description: Posts sorted by publication time, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Corpus read failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/blogs/{blogId}:
    delete:
      summary: Deactivate a registered blog without deleting its corpus
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/BlogId"
      responses:
        "200":
          description: Blog deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid blog id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Blog not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/blogs/{blogId}/refresh:
    post:
      summary: Start an incremental crawl for one registered blog
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/BlogId"
      responses:
        "202":
          description: Incremental crawl request result
          content:
            application/json:
              schema:
                type: object
                required: [ok, job]
                properties:
                  ok:
                    type: boolean
                    const: true
                  job:
                    $ref: "#/components/schemas/BlogJobStart"
        "400":
          description: Invalid blog id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/blogs/{blogId}/posts:
    get:
      summary: List posts for one registered blog
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/BlogId"
        - $ref: "#/components/parameters/BlogPostOffset"
        - $ref: "#/components/parameters/BlogPostLimit"
      responses:
        "200":
          description: Posts sorted by publication time, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostsResponse"
        "400":
          description: Invalid blog id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Corpus read failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/blogs/{blogId}/posts/{articleId}:
    get:
      summary: Read one crawled Naver blog article
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/BlogId"
        - $ref: "#/components/parameters/BlogArticleId"
      responses:
        "200":
          description: Article markdown and crawl metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogArticleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Article not found or path invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: attn_session
  parameters:
...
    BlogId:
      name: blogId
      in: path
      required: true
      schema:
        type: string
        minLength: 2
        maxLength: 30
        pattern: "^[a-zA-Z0-9_-]{2,30}$"
    BlogArticleId:
      name: articleId
      in: path
      required: true
      schema:
        type: string
        pattern: "^naver-[a-zA-Z0-9_-]{2,30}-[0-9]+$"
    BlogPostOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    BlogPostLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    Unauthorized:
      description: Login required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    BlogListResponse:
      type: object
      required: [ok, blogs]
      properties:
        ok:
          type: boolean
          const: true
        blogs:
          type: array
          items:
            $ref: "#/components/schemas/BlogListItem"
    BlogListItem:
      type: object
      required: [id, name, source, tags, active, addedAt, count, recentCount]
      properties:
        id:
          type: string
          pattern: "^[a-zA-Z0-9_-]{2,30}$"
        name:
          type: string
        source:
          type: string
          enum: [naver]
        tags:
          type: array
          items:
            type: string
        active:
          type: boolean
        addedAt:
          type: string
          format: date-time
        crawlSince:
          type: string
          format: date-time
          description: Earliest Naver post publication time eligible for collection.
        count:
          type: integer
          minimum: 0
        recentCount:
          type: integer
          minimum: 0
        fetchBlocked:
          type: boolean
        latest:
          oneOf:
            - $ref: "#/components/schemas/BlogLatestPost"
            - type: "null"
        lastCheck:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        runningJob:
          oneOf:
            - $ref: "#/components/schemas/BlogRunningJob"
            - type: "null"
        lastJob:
          oneOf:
            - $ref: "#/components/schemas/BlogLastJob"
            - type: "null"
    BlogLatestPost:
      type: object
      required: [id, title, url, publishedAtText, fetchedAt]
      properties:
        id:
          type: string
        title:
          type: string
        url:
          type: string
        publishedAtText:
          type: string
        fetchedAt:
          type: string
    BlogRunningJob:
      type: object
      required: [mode, startedAt]
      properties:
        mode:
          type: string
          enum: [backfill, refresh]
        startedAt:
          type: string
          format: date-time
    BlogLastJob:
      type: object
      required: [discovered, saved, failed, done, updatedAt]
      properties:
        discovered:
          type: integer
          minimum: 0
        saved:
          type: integer
          minimum: 0
        failed:
          type: integer
          minimum: 0
        done:
          type: boolean
        updatedAt:
          type: string
    BlogCreateRequest:
      type: object
      additionalProperties: false
      required: [blogId]
      properties:
        blogId:
          type: string
          minLength: 2
          maxLength: 30
          pattern: "^[a-zA-Z0-9_-]{2,30}$"
        name:
          type: string
        crawlSince:
          type: string
          format: date-time
          description: Optional lower publication-time boundary for the initial crawl and later refreshes.
    BlogCreateResponse:
      type: object
      required: [ok, blog, preview, job]
      properties:
        ok:
          type: boolean
          const: true
        blog:
          $ref: "#/components/schemas/BlogRegistryItem"
        preview:
          $ref: "#/components/schemas/BlogPreview"
        job:
          $ref: "#/components/schemas/BlogJobStart"
    BlogRegistryItem:
      type: object
      required: [id, name, source, tags, active, addedAt]
      properties:
        id:
          type: string
        name:
          type: string
        source:
          type: string
          enum: [naver]
        tags:
          type: array
          items:
            type: string
        active:
          type: boolean
        addedAt:
          type: string
          format: date-time
        crawlSince:
          type: string
          format: date-time
          description: Earliest Naver post publication time eligible for collection.
    BlogPreview:
      type: object
      required: [totalCount, items]
      properties:
        totalCount:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            type: object
            required: [logNo, title, nickName]
            properties:
              logNo:
                type: string
              title:
                type: string
              nickName:
                type: string
    BlogJobStart:
      type: object
      required: [started]
      properties:
        started:
          type: boolean
        reason:
          type: string
          enum: [already-running]
    BlogPostsResponse:
      type: object
      required: [ok, total, posts]
      properties:
        ok:
          type: boolean
          const: true
        total:
          type: integer
          minimum: 0
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
    BlogPost:
      type: object
      required: [id, blogId]
      properties:
        id:
          type: string
        blogId:
          type: string
        title:
          type: string
        author:
          type: string
        source:
          type: string
        url:
          type: string
        publishedAtText:
          type: string
        publishedAt:
          type: string
        fetchedAt:
          type: string
        contentHash:
          type: string
        markdownPath:
          type: string
        metadataPath:
          type: string
        imageCount:
          type: integer
          minimum: 0
        summary:
          type: [string, "null"]
        summaryReason:
          type: [string, "null"]
    BlogArticleResponse:
      type: object
      required: [ok, markdown, metadata]
      properties:
        ok:
          type: boolean
          const: true
        markdown:
          type: string
        metadata:
          oneOf:
            - $ref: "#/components/schemas/BlogArticleMetadata"
            - type: "null"
    BlogArticleMetadata:
      type: object
      required: [id, source, author, blogId, logNo, title, url, fetchedAt, contentHash]
      properties:
        id:
          type: string
        source:
          type: string
          enum: [naver_blog]
        sourceReliability:
          type: string
          enum: [scraped]
        ingestMode:
          type: string
          enum: [crawl]
        fetchStatus:
          type: string
          enum: [ok]
        author:
          type: string
        blogId:
          type: string
        logNo:
          type: string
        title:
          type: string
        category:
          type: string
        url:
          type: string
        canonicalUrl:
          type: string
        publishedAtText:
          type: string
        fetchedAt:
          type: string
          format: date-time
        contentHash:
          type: string
        rawHtmlPath:
          type: string
        markdownPath:
          type: string
        imageCount:
          type: integer
          minimum: 0
        imageDownloadEnabled:
          type: boolean
        images:
          type: array
          items:
            type: object
            additionalProperties: true
        listItem:
          type: object
          additionalProperties: true
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
        error:
          type: string
...